+++ History.md
@@ -1,4 +1,12 @@
-0.2.0 / 2012-12-14
+0.2.1 / 2012-12-18
+==================
+
+* Added the `pageview` method for tracking virtual pageviews
+* Added Travis-CI
+* Fixed window level objects in customerio and gosquared
+
+
+0.2.0 / 2012-12-16
 ==================
 
 * Separated providers into separate files for easier maintenance
@@ -6,8 +14,6 @@
 * Moved `utils` directly onto the analytics object
 * Added `extend` and `alias` utils
 * Added `settings` defaults for all providers
-* Added Travis-CI
-* Fixed window level objects in customerio and gosquared
 
 
 0.1.2 / 2012-12-14